+++ source/objects/randomgenerator/wrap_randomgenerator.cpp
@@ -7,21 +7,6 @@
 #define CLASS_TYPE LUAOBJ_TYPE_RANDOMGENERATOR
 #define CLASS_NAME "RandomGenerator"
 
-#include <limits>
-
-template <typename T>
-static T checkrandomseed_part(lua_State *L, int idx)
-{
-    double num = luaL_checknumber(L, idx);
-    double inf = std::numeric_limits<double>::infinity();
-
-    // Disallow conversions from infinity and NaN.
-    if (num == inf || num == -inf || num != num)
-        luaL_argerror(L, idx, "invalid random seed");
-
-    return (T) num;
-}
-
 int randomgeneratorNew(lua_State * L)
 {
     RandomGenerator::Seed s;
